Reviewed-on: https://01.kood.tech/git/Jacob/multiplayer-fps/pulls/92

+++ common/src/defaults.rs
@@ -12,8 +12,8 @@ pub const MAP_HEIGHT: usize = 21; // Must be odd
 pub const MAP_BRANCHING: f32 = 0.5; // 0.0..=1.0
 pub const MAP_OPENNESS: f32 = 0.3; // 0.0..=1.0
 
-pub const PLAYER_MAX_HP: u32 = 100;
-pub const DEFAULT_PLAYER_HP: u32 = PLAYER_MAX_HP;
+pub const PLAYER_MAX_HP: f32 = 100.0;
+pub const DEFAULT_PLAYER_HP: f32 = PLAYER_MAX_HP;
 pub const DEFAULT_PLAYER_NAME: &str = "Player";
 pub const PLAYER_SPEED: f32 = 0.1;
 pub const PLAYER_SIZE: f32 = 0.25;
